Remove `tmp::entry`
