--- include/tmp/entry
@@ -1,65 +1,0 @@
-#ifndef TMP_ENTRY_H
-#define TMP_ENTRY_H
-
-#include <tmp/export>
-
-#include <cstddef>
-#include <filesystem>
-
-namespace tmp {
-
-/// `tmp::entry` is a smart handle that owns and manages a temporary path and
-/// deletes it recursively when this handle goes out of scope
-///
-/// The managed path is deleted of when either of the following happens:
-/// - the managing `tmp::entry` object is destroyed
-/// - the managing `tmp::entry` object is assigned another path via `operator=`
-///
-/// Subclasses should provide a path to manage for this class;
-/// any opening or closing operations should be managed by subclasses
-class TMP_EXPORT entry {
-public:
-  /// Returns the managed path
-  /// @returns The full path this entry manages
-  operator const std::filesystem::path&() const noexcept;
-
-  /// Returns the managed path
-  /// @returns The full path this entry manages
-  const std::filesystem::path& path() const noexcept;
-
-  /// Deletes the managed path recursively
-  virtual ~entry() noexcept;
-
-  entry(const entry&) = delete;               ///< not CopyConstructible
-  entry& operator=(const entry&) = delete;    ///< not CopyAssignable
-
-  bool operator==(const entry& rhs) const noexcept;    ///< EqualityComparable
-  bool operator!=(const entry& rhs) const noexcept;
-  bool operator<(const entry& rhs) const noexcept;    ///< LessThanComparable
-  bool operator>(const entry& rhs) const noexcept;
-  bool operator<=(const entry& rhs) const noexcept;
-  bool operator>=(const entry& rhs) const noexcept;
-
-protected:
-  /// Creates a temporary entry which owns the given path
-  /// @param path A path to manage
-  explicit entry(std::filesystem::path path) noexcept;
-
-  /// Clears the stored pathname
-  void clear() noexcept;
-
-  entry(entry&&) noexcept;               ///< MoveConstructible
-  entry& operator=(entry&&) noexcept;    ///< MoveAssignable
-
-private:
-  /// The managed path
-  std::filesystem::path pathobject;
-};
-}    // namespace tmp
-
-/// The template specialization of `std::hash` for `tmp::entry`
-template<> struct TMP_EXPORT std::hash<tmp::entry> {
-  std::size_t operator()(const tmp::entry& entry) const noexcept;
-};
-
-#endif    // TMP_ENTRY_H
